chore: ssr-debug

+++ src/server/graphql/schema.ts
@@ -20,12 +20,13 @@ if (env.APP_ENV === 'local') {
     const graphql = await import('graphql');
     const schemaAsString = graphql.printSchema(graphql.lexicographicSortSchema(schema));
     const logger = getLogger('schema');
-    const filePath = path.join(__dirname, '../../../../src/server/graphql/schema.graphql');
-    await fs.writeFile(filePath, schemaAsString);
-    logger.trace(`Schema file written to: ${filePath}`);
+    const filePath = path.join(__dirname, '../../../../../src/server/graphql/schema.graphql');
+    const existingContent = await fs.readFile(filePath, 'utf-8');
+    if (existingContent !== schemaAsString) {
+      await fs.writeFile(filePath, schemaAsString);
+      logger.debug(`Schema file written to: ${filePath}`);
+    }
   } catch (error) {
     logger.error('Error writing schema file:', error);
   }
-} else {
-  logger.debug('Skipping file writing in non-local environment.');
 }
